Python_Code_3
Testing of a simple code that accepts user input and summarizes sales by a given sort field and displays the results to the user.
code must prompt user to input 'gender' or 'state'. If value is invalid, error message must be displayed to prompt user to re-enter value. Code also involves reading of json file.